add media parent class
